completed edit spot but didn't create middleware for authorization. Created conditional instead

+++ backend/db/migrations/20220701233427-create-spot.js
@@ -33,11 +33,11 @@ module.exports = {
         allowNull: false
       },
       latitude: {
-        type: Sequelize.FLOAT(10, 7),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       longitude: {
-        type: Sequelize.FLOAT(10, 7),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       name: {
